requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-openai
langgraph
python-dotenv
requests
pydantic
mcp[cli]
google-api-python-client
google-auth